Import Cloud : fix acces budgea pour activation/desactivation des comptes

+++ budgetview/bv_desktop/src/main/java/com/budgetview/desktop/importer/components/CloudConnectionDateFieldEditor.java
@@ -33,7 +33,7 @@ public class CloudConnectionDateFieldEditor extends CloudConnectionFieldEditor {
           this.setText("");
         }
         else {
-          this.setText(Lang.get("month." + ((Integer) value) + ".long"));
+          this.setText(Lang.get("month." + value + ".long"));
         }
         return this;
       }
